update scripts for package
