+++ .env
@@ -1,2 +1,2 @@
 HTTP_PORT=8181
-WEB_SOCKET_PORT=4000
+WEB_SOCKET_PORT=3000
